refactor(vscode): lift history timeline + row construction into historyTimeline.ts
Same playbook as #814 / #815 / #816 (cardBuilder lifecycle lifts) for
the History panel. Lifts the timeline + per-row DOM operations out of
`history/behavior.ts` into a new sibling `historyTimeline.ts`. The
diff sub-views (`fillDiff`, `computeDiffStats`, `applyDiffStats`,
`renderHistoryDiffMode`, `buildHistoryDiffStack`, `buildDiffPane`,
`showDiff`) are a separate concern with their own state — they'll be
a follow-up PR.

New `historyTimeline.ts` exports:
- `renderTimeline(entries, config)` — rebuild the whole timeline.
  Iterates over the manifest, creates a `<div class="row">` per
  entry, wires up the click/diff/shift-click handlers.
- `toggleSelected(id, row, config)` — max-2 selection set used for
  the toolbar diff button.
- `expandRow(id, row, config)` — toggle the inline `.expanded`
  full-image view; posts `loadImage` so the extension streams bytes.
- `requestRowDiff(id, row, against, config)` — open an inline diff
  expansion (vs previous / vs current); posts `requestDiff`.
- `fillExpansion(id, imageData, entry, config)` — populate an open
  expansion with the streamed PNG + an "Open in Editor" action.
- `populateThumb(thumbEl, imageData)` — pure DOM helper for the lazy
  thumb column.

`HistoryRowConfig` covers the collaborator surface: `vscode`,
`timelineEl`, `btnDiffEl`, `selectedIds` Set, `getExpandedId()` /
`setExpandedId()` accessors over the mutable scalar, `thumbCache`
Map, `thumbRequested` Set, `thumbObserver`. The state stays owned by
`behavior.ts` so the lift is purely mechanical.

`behavior.ts` builds the `rowConfig` once at panel boot and exposes
a thin `renderTimeline()` shim that closes over `entries` for the
two `setEntries` / `entryAdded` call sites. The `imageReady` handler
calls `fillExpansionDom` directly with the config; `thumbReady` calls
the imported `populateThumb` (no config needed).

## Drive-by

Drops four now-unused imports from `behavior.ts`:
- `escapeHtml`, `findLatestMainHash`, `formatAbsolute`, `formatRelative`
  from `./historyData` — all four were only referenced from the
  lifted timeline functions.

`history/behavior.ts`: 751 → 579 lines (−172 LOC).
`historyTimeline.ts`: +269 lines (new file).

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 373/373 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke: open a History panel scope:
  - `setEntries` populates the timeline; thumbs lazy-load via the
    `IntersectionObserver`.
  - Clicking a row expands the full image (`imageReady` arrives,
    `fillExpansionDom` paints the `<img>` + "Open in Editor" button).
  - Click again collapses.
  - Shift+click two rows + toolbar Diff — toolbar diff button enables
    only with exactly 2 selected.
  - Diff vs previous / vs current per-row buttons open an inline
    diff expansion.
